requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
pillow>=9.0.0
numpy>=1.24.0
opencv-python>=4.8.0
huggingface-hub>=0.16.0
accelerate>=0.20.0
safetensors>=0.3.0
timm>=0.9.0
einops>=0.7.0
